fix: do not cache io errors (#5856)
When we compile & cache the contract, there are two very different kinds
of "errors" we can have:

* IO error due to the database being corrupted and such
* Contract failed to compile because it contained invalid WASM

The second outcome is normal, expected mode of execution and should be
cached. The first error is a genuine exceptional situation, and should
not be cached (eg, retrying the operation might succeed).


Test Plan
=======

Existing tests to verify that the happy path still works. I don't believe we current have fault injection tests to make sure genuine IO-errors are handled correctly.