feat: Don't hide errors with parsing `network config`. (#5705)
`client config` contains a sections for `near-network` called `NearNetworkConfig`.

In function `validate()` of `NearNetworkConfig` we verify all argument. However, if any argument is invalid, we will print an error an continue running an normal. 

The proper solution is to, throw an error, and make let the caller know, that the `network_config` is invalid. Btw, we don't check whenever the number of peers allowed is to high. See https://github.com/near/nearcore/pull/5703