client: add GetBlockHash request and use it in Rosetta RPC (#6238)
The `/network/status` Rosetta RPC returns, among other things, height
and hash of the genesis block.  Currently, it does it by fetching the
genesis block.  Sadly, in non-archival nodes, genesis block is usually
unavailable which leads to failures in serving the RPC.  To address
this issue, introduce a `GetBlockHash` request which returns only the
hash of a block.  Since height→hash mapping is not garbage collected,
this request will always work even for garbage collected blocks.