neard: generate NEARD_VERSION et al via `build.rs` script (#6188)
On it’s own this change doesn’t really offer us a lot.  There is
a tiny optimisation where some values don’t need to be constructed
lazily at run time but are embedded into the executable but that’s
really not an important win.

The actual reason for this change is to later extend the `build.rs`
script with other information.  Namely, rustc version is one such
piece of information that we’ll like to include.

Issue: https://github.com/near/nearcore/issues/5979