Chunk delay metric: report delay per shard (#6381)
Adds a new metric by tracking the event of requesting a chunk.

Changes how the timestamps are stored. Instead of triggering metric update at every opportunity, wait for the block to be processed and then update the metrics. Makes things a lot simpler.


Updates the existing chunks delay metric to report a histogram:
* Time between learning about a block and receiving the last of its chunks

Metric "blocks ahead of head" is removed as redundant, as it is mostly the same as the number of orphan blocks.

Dump of metrics from a betanet node with this PR: https://gist.github.com/nikurt/975dd6b33fe0361ab1f918515c8526f1