chain: rosetta: guarantee order of operations in transactions (#6193)
Because we are using a hash set when identifying accounts which were
changed in a block, the order in which change events are collected is
nondeterministic.  This means that operations in transactions which
cause the changes in nondeterministic order as well.

This is an issue because we assign operation indexes based position in
the list.  This means, that querying the same block multiple times may
result in transactions whose operation indexing changes.

To make sure that doesn’t happen and operation indexes remain
consistent store accounts in a vector in the same order the view
client returns them.