Register the subscriber globally (#6506)
* Register the subscriber globally

When the subscriber is registered globally, it transparently also
applies to any thread, unlike the subscribers that are registered using
`set_default`, which only apply to the current thread.

This is somewhat incompatible with the non-blocking writer, because now
it is possible that we drop the writer guard in the main thread, flush
the `stderr` and then threads log additional messages before the program
exits. This would then lead to some messages being lost.

I feel like we can afford this, though. The alternative would be to
somehow ensure we register the subscriber for every thread, but thread
creation is not necessarily have a full control over. Libraries could
spawn their own, etc.

* Move `dead_core` attribute to a field

Co-authored-by: near-bulldozer[bot] <73298989+near-bulldozer[bot]@users.noreply.github.com>