Switch __count field in Enum to strum (#6526)
fix: Use strum crate for consistency

This PR is for problem #6520. It replaces ```__count``` field in ```ExtCosts``` & ```ActionCosts``` Enums with ```strum::EnumCount``` and change codes where count of fields in Enum is needed. As we used strum crates in a few place in our project, this change is meant for consistency.

Compilation and existing unit test or ```core/primitives-core``` passed