nearcore: take max_open_files into consideration in recompress_storage (#6612)
So far recompress_storage assumed that a single database will use 512
file descriptors at the most and thus set NOFILE limit to three times
that (so it can open two databases simultaneously and have whole bunch
of descriptors to spare).  However, this number is now configurable
and may be larger.  Change the command to take into consideration what
is the actual configured max_open_files number.