store: change max_open_files to u32 (#6613)
Change type of max_open_files configurable setting from i32 to u32
so that we don’t have to deal with negative values.  For example,
we don’t need to worry about `max_open_files as u64` blowing up to
astronomical values.  Putting a negative value for the field is not
valid anyway.