Merge pull request from GHSA-5vx4-7hvq-wr66
* Move function count validation to wasmparser

* Limit the number of locals in the contract

The large number of locals in WASM can be encoded in a constant amount
of space. It is straightforward for a piece of code operate with these
locals in a poor way – in fact we do so ourselves in multiple places,
expanding the constant-length encoding a linear vector containing the
local types.

While this new representation gives a `O(1)` access to any local's type,
constructing the vector like this in the first place is a linear-time
ordeal, one that the attacker can control by changing the value of a
couple of bytes.

Today we only have a limit (and a pretty generous one) per-function
conetained within the contract. Given a typical function limit (`10000`)
and number of locals per function imposed by `wasmparser` (`50000`), we
would end up up constructing vectors for up-to 500M local types,
possibly multiple times over, and most of the time with some
reallocations in between to resize the vectors. That's a lot of work!

In practice our contracts are also limited in size to 4MiB. Each
instruction to access (`local.get`, `local.set` or `local.tee`) is at
least 2 bytes and must specify a constant local index. This means that
an useful number of locals is largely limited by the size of the
contract. Even a contract which only ever reads and writes locals will
need at least 2 instructions for each “useful” local, so 1M of locals
seems like a very reasonable contract-wide upper limit on the number of
locals to impose.