refactor: Clean up VMResult::NotRun (#6673)
Replace results without outcome with no-op outcome results.

Note that `VMOutcome` itself cannot be no-op, it needs to be considered
in conjunction with the error. Therefore, the constructor is on the
`VMResult` and requires an error.
If we move the error to a field inside `VMOutcome`, this will also
become clearer.