feat: add runtime checks for "immutable" columns (#6670)
We have three classes of columns in the DB:

* read-write columns which are modified arbitrary
* insert only columns, where each value is written once and exists
  unchanged until GCed. You can call such columns immutable.
* RC columns, which are also immutable, but allow repated writes which
  bump rc.

Today, we use the same operations for working with both read-write
and write-once columns. In this PR, I want to seprate that into two
semantically distinct operations (which would allow us to have better
runtime and compile-time checks):

* `set` for read-write columns
* `insert` for insert-only columns

The main motivation is to make reasoning about the code easier -- if a
bit of code only does inserts and no sets, it can be reorderd relatively
easily.