bugfix: fix m1 sandbox (#6544)
We have not implemented build caching for wasmtime yet. As such, we also
need to not try using it, or the build will fail.

Partial fix for https://github.com/near/workspaces-rs/issues/110

@matklad I'm not sure how good I feel about exposing VMKind while it was previously marked as internal. It was already being used by at least near-vm-runner-standalone and runtime-param-estimator, but still. The issue is, if I'm not doing that it means that somewhere else down the call chain we would have to ignore a correctly-configured contract cache, which feels worse to me than exposing VMKind.

What do you think about that? Or is internal maybe for "anything in nearcore but semver doesn't apply"?

cc @ChaoticTempest