test: Snapshot tests for estimator warehouse (#6780)
Restyling tests for commands to be more e2e and more data-driven.
The insta crate is used, since it makes life easier and it is already
used in other parts of the code base.

Previous tests were trying to cover all pub(crate) methods, which is not
the best style for us. Instead, use input and output strings that are
independent of the implementation.