Consistently use spans to instrument important node functions (#6777)
Make logging more useful for viewing tracing in a distributed tracing system, such as Jaeger.

The goal is to group spans together if they are triggered by a single event. For example, calling `check_triggers()` and processing accepted blocks is triggered by a single network message. Another example is parallel processing of chunks, it's much better to track them as children of a parent span.

These changes make logging more verbose by 20% in terms of the number of lines. Measured by running a betanet node for 5 minutes.