chain-primitives: use StorageError kind and avoid string allocation (#6773)
Firstly, in runtime use StorageError kind for converting StorageError
errors to Error rather than formatting them to string and treating
them as Other.

Secondly, self.get_epoch_and_valset in tests already returns an Error
so there’s no need to map it to anything.

Lastly, get rid of `From<String>` implementation and in the one place
it was needed (`gc_col_block_per_height`) use a closure with explicit
ErrorKind::Other construction.  This gets rid of a string allocation
which used to always happen; now it only happens when error is
returned.