store: prefer enum_map::EnumMap to strum’s count and iterator (#6829)
RocksDB and TestDB both create vectors which are meant to map from
column to some value.  The issue is that the way they are created
assumes that variants in DBCol are all listed in order of their
values, the values are all unique and there are no holes.  In other
words, removing or rearranging variants of DBCol causes things to
break.

Solve that by using enum_map::EnumMap which creates fixed-size
constant-access arrays (same as a Vec approach) but allow the map to
be indexed by the enum value itself which avoids all of the above
issues.

Furthermore, introduce RocksDB::cf_handle method which consolidates
all bunch of unsafe blocks into a single unsafe block in that method.
Having a single unsafe block makes it easier to reason about the
safety guarantees.