fix: reduce spammy logging (#6880)
I've noticed that `check_triggers` accounts for a high precentage of
DEBUG-level logging. This is because we check very frequently, but often
the check is no-op, as we don't have to run any triggers. So remove the
span from checking, and make sure that we always span the actual
triggers themselves.

That is, this should now log only when the trigger is executed