Run mocknet nodes on port 3030 (#6883)
Mocknet scripts assume the nodes run on port 3030, for example here https://github.com/near/nearcore/blob/master/pytest/lib/mocknet_helpers.py#L10, but it's not the only reference to port 3030.
After a recent infrastructure change the nodes run on port 4040 by default.

This PR sets a port that works well with mocknet. An alternative would be to update mocknet scripts to use a different port, but that requires more maintenance.