store: use match instead of col_set for DBCol::if_foo methods (#6951)
Eliminate custom col_set function (which creates a mapping from all
DBCol variants to a boolean value) in favour of using match statement
in methods which used those mappings.

The main motivation for this is to get rid of the dependency on values
assigned to DBCol variants and the strum::EnumCount derive.

The problem is that strum::EnumCount defines COUNT constant which
returns number of variants while what we actually want is the value
one greater than the highest value of any variant.  This is currently
the same, but as we might want to delete deprecated variants those two
numbers would diverge.