Add check in preprocess_block to check no challenged block on chain (#7075)
In postprocess_block, when it updated tip, it checks that no challenged block is on chain. This check should be performed in preprocess_block instead of postprocess_block. After this change, postprocess_block shouldn't return errors other than database errors. This change will make the error handling of of postprocess_block #6968 easier.