store: separate StoreUpdate::update_refcount into increase and decrease (#7097)
Every time StoreUpdate::update_refcount is used it’s called with
a positive or a negative value.  I.e. each caller knows whether they
are increasing or decreasing the value.  This makes it trivial to
split the method into increase_refcount and decrease_refcount methods.
The two methods are now simpler since they don’t have to check sign of
the delta argument.