feat: record detailed trace of IO operations (#7052)
Allows both neard and the runimte-params-estimator to collect a
human-readable trace of executed storage and DB operations.

The tracing infrastructure is used here to collect spans and events
at runtime. The new subscriber layer IoTraceLayer can be enabled
for neard and for the runtime estimator, with the flag
--record-io-trace=my_file.log.

A new tracing target "io_tracer" has been added for events that are
only of interest to this new subscriber layer. And all such event are
emitted on the Trace level.

To avoid affecting release mode neard, I added the feature
"io_trace" to store and vm-logic.

Some existing spans are also reused, those are in the targets
"host-function", "store", "vm_logic", "runtime".

The recorded trace can be used to analyse DB access patterns.
Emphasis on *patterns*, as the timing of IO events is not
observable in the trace.

In a next step, the estimator could use such a trace to estimate
IO costs indirectly.

Resolves #7053, which has examples of the output format.