network: bypass near_store::Store abstraction in network’s Store (#7149)
Motivation for this is that there are incoming many intrusive changes
to the near_store::Store interface.  If network’s Store no longer uses
near_store::Store and goes through near_store::db::Database directly
changes to the former will not affect it.

On the flip side it does expose more of the near_store::db module.