network: get rid of RoutedMessageBody::Query{Request,Response} (#7282)
We haven’t been sending those messages for a while now so support for
borsh serialising and deserialising them is unnecessary.  Meanwhile,
the serialisation forces bunch of other types to be borsh-serialisable
as well making it harder to change those types.  And changing them is
something we might want to do.

The message used to be sent when view client couldn’t handle a Query
due to lack of state.  In that case it would route the message to
other nodes.  This feature has been removed in commit f68140e65c48:
‘Add structured errors to query method in ViewClient and RPC.’