Simplify DBCol’s Display implementation and prefer it over Debug (#7399)
I’d argue that the current DBCol’s Display implementation offers no
value.  I cannot think of a situation in which we’d rather see the
description of the column provided by the implementation rather than
the variant name.

If one sees the name of the column in logs or some other places, it’s
trivial to start grepping the source code for uses of that column.
When the description is shown it first need to be transformed back
into the variant name to actually be useful.

And it’s not like the description explains anything.  For example,
‘epoch information’ is no more clear than EpochInfo.  In a few cases,
things like ‘hash of chunk per height and shard_id’ are easier to read
than ChunkPerHeightShard but I’d still question if this is enough to
justify existence of this custom implementation.

In this change I suggest 1) replacing Display implementation by one
which simply outputs the variant name and 2) converting existing code
to prefer using Display rather than Debug formatting for columns.

With that we get rid of 50 lines of code, will never again need to
look at the Display implementation (e.g. when adding or removing
columns) and make it simpler to format columns with plain `{col}`.

Note that the Debug derive needs to remain as it is (indirectly) used
in a couple of asserts and tests.