fix: fix logical deadlock in the network  (#7318)
We should send ChainInfo to net whenever it changes, but we've missed
one case: at the startup, theres' a change due to us getting the initial
info. If we don't do that, we can't start block processing, so the info
is never updated. So, let's just send the info on startup.

There's a python tests which uncovered this bug:

   $ python3 tests/sanity/block_sync_archival.py


While fixing this, I've also noticed a bug in the epoch manager, which failed to report validators for genesis block, so that is fixed as well. 

best reviewed per commit.